chore: refactor providers
