+++ src/events.ts
@@ -1,21 +1,16 @@
-import fetch from 'node-fetch';
 import snapshot from '@snapshot-labs/snapshot.js';
-import { sendEventToDiscordSubscribers } from './discord';
-import { sendPushNotification } from './helpers/beams';
-import db from './helpers/mysql';
-import { sha256 } from './helpers/utils';
-import { getProposal } from './helpers/proposal';
 import { capture } from '@snapshot-labs/snapshot-sentry';
-import { timeOutgoingRequest } from './helpers/metrics';
+import db from './helpers/mysql';
+import { getProposal, getSubscribers } from './helpers/utils';
+import { send as sendDiscord } from './providers/discord';
+import { send as sendBeams } from './providers/beams';
+import { send as sendWebhook } from './providers/webhook';
 
-const delay = 5;
-const interval = 15;
-const HTTP_WEBHOOK_TIMEOUT = 15000;
-const serviceEvents = parseInt(process.env.SERVICE_EVENTS || '0');
-const serviceEventsSalt = parseInt(process.env.SERVICE_EVENTS_SALT || '12345');
-const servicePushNotifications = parseInt(process.env.SERVICE_PUSH_NOTIFICATIONS || '0');
+const DELAY = 5;
+const INTERVAL = 15;
+const SERVICE_EVENTS = parseInt(process.env.SERVICE_EVENTS || '0');
 
-export const handleCreatedEvent = async event => {
+export async function handleCreatedEvent(event) {
   const { space, id } = event;
   const proposalId = id.replace('proposal/', '') || '';
   const proposal = await getProposal(proposalId);
@@ -52,9 +47,9 @@ export const handleCreatedEvent = async event => {
     });
   }
   return db.queryAsync(query, params);
-};
+}
 
-export const handleDeletedEvent = async event => {
+export async function handleDeletedEvent(event) {
   const { ipfs } = event;
   const ipfsData = await snapshot.utils.ipfsGet('pineapple.fyi', ipfs);
   const proposalId = ipfsData.data.message.proposal;
@@ -68,63 +63,27 @@ export const handleDeletedEvent = async event => {
     INSERT IGNORE INTO events SET ?;
   `;
   return db.queryAsync(query, [event.id, event]);
-};
-
-export async function sendEvent(event, to, method = 'POST') {
-  event.token = sha256(`${to}${serviceEventsSalt}`);
-  event.secret = sha256(`${to}${serviceEventsSalt}`);
-  const headerSecret = sha256(`${to}${process.env.SERVICE_EVENTS_SALT}`);
-  const url = to.replace('[PROPOSAL-ID]', event.id.split('/')[1]);
-  const end = timeOutgoingRequest.startTimer({ method });
-  let res;
-
-  try {
-    res = await fetch(url, {
-      method,
-      headers: {
-        'Content-Type': 'application/json',
-        Authentication: headerSecret
-      },
-      body: JSON.stringify(event),
-      timeout: HTTP_WEBHOOK_TIMEOUT
-    });
-    return res.text();
-  } catch (error: any) {
-    if (error.message.includes('network timeout')) {
-      console.error('[events] Timed out while sending the webhook', url);
-    } else {
-      console.error('[events] Error sending event data to webhook', url, JSON.stringify(error));
-    }
-    throw error;
-  } finally {
-    end({ status: res?.statusCode || 0 });
-  }
 }
 
-const sendEventToWebhookSubscribers = (event, subscribers) => {
-  Promise.allSettled(
-    subscribers
-      .filter(subscriber => [event.space, '*'].includes(subscriber.space))
-      .map(subscriber => sendEvent(event, subscriber.url, subscriber.method))
-  )
-    .then(() => console.log('[events] Process event done'))
-    .catch(e => capture(e));
-};
+async function processEvents() {
+  const ts = ~~(Date.now() / 1e3) - DELAY;
 
-async function processEvents(subscribers) {
-  const ts = parseInt((Date.now() / 1e3).toFixed()) - delay;
   const events = await db.queryAsync('SELECT * FROM events WHERE expire <= ?', [ts]);
-
   console.log('[events] Process event start', ts, events.length);
 
   for (const event of events) {
     const proposalId = event.id.replace('proposal/', '');
+    const proposal = await getProposal(proposalId);
+    const subscribers = await getSubscribers(event.space);
 
-    // Send event to discord subscribers and webhook subscribers and then delete event from db
-    // TODO: handle errors and retry
-    if (servicePushNotifications && event.event === 'proposal/start') sendPushNotification(event);
-    sendEventToDiscordSubscribers(event.event, proposalId);
-    sendEventToWebhookSubscribers(event, subscribers);
+    if (proposal) {
+      // TODO: handle errors and retry
+      sendBeams(event, proposal, subscribers);
+      sendDiscord(event, proposal, subscribers);
+      sendWebhook(event, proposal, subscribers);
+    } else {
+      console.log(`Proposal ${proposalId} not found`);
+    }
 
     try {
       await db.queryAsync('DELETE FROM events WHERE id = ? AND event = ? LIMIT 1', [
@@ -141,15 +100,16 @@ async function processEvents(subscribers) {
 
 async function run() {
   try {
-    const subscribers = await db.queryAsync('SELECT * FROM subscribers');
-    console.log('[events] Subscribers', subscribers.length);
-    await processEvents(subscribers);
+    await processEvents();
   } catch (e) {
     capture(e);
     console.log('[events] Failed to process', e);
   }
-  await snapshot.utils.sleep(interval * 1e3);
+
+  await snapshot.utils.sleep(INTERVAL * 1e3);
   await run();
 }
 
-if (serviceEvents) setTimeout(() => run(), interval * 1e3);
+if (SERVICE_EVENTS) {
+  setTimeout(() => run(), INTERVAL * 1e3);
+}
